Create base ethernet class `Ethernet::IdfService`
