+++ Sming/Components/Network/src/Platform/Ethernet.h
@@ -61,14 +61,6 @@ constexpr int8_t PIN_DEFAULT{-2};
 constexpr int8_t PIN_UNUSED{-1};
 
 /**
- * @brief Configuration for Ethernet MAC
- */
-struct MacConfig {
-	int8_t smiMdcPin = PIN_DEFAULT;  //< SMI MDC GPIO number
-	int8_t smiMdioPin = PIN_DEFAULT; //< SMI MDIO GPIO number
-};
-
-/**
  * @brief Link speed
  */
 enum class Speed {
@@ -119,14 +111,6 @@ public:
 };
 
 /**
- * @brief Service configuration options
- */
-struct Config {
-	MacConfig mac;
-	PhyConfig phy;
-};
-
-/**
  * @brief Abstract Service class
  *
  * Provides a common implementation for TCP/IP ethernet support.
@@ -143,14 +127,6 @@ struct Config {
 class Service
 {
 public:
-	/**
-	 * @brief Configure and start the ethernet service
-	 * @param config Configuration options
-	 *
-	 * Applications should expect to receive Start and Connected events following this call.
-	 */
-	virtual bool begin(const Config& config) = 0;
-
 	/**
 	 * @brief Tear down the ethernet connection
 	 */
